criando exemplo de autenticação

+++ app/Http/Controllers/Api/SeriesController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\SeriesFormRequest;
 use App\Models\Series;
 use App\Repositories\SeriesRepository;
+use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Http\Request;
 
 class SeriesController extends Controller
@@ -82,9 +83,13 @@ class SeriesController extends Controller
      * @param  int $series
      * @return Response
      */
-    public function destroy (int $series)
+    public function destroy (int $series, Authenticatable $user)
     {
-        Series::destroy($series);
-        return response('Série excluida com sucesso',200);
+        if($user->tokenCan('is_admin')){
+            Series::destroy($series);
+            return response('Série excluida com sucesso',200);
+        }
+
+        return response()->json(["mensagem"=>"Ação não autorizada", "status"=> 401], 401);
     }
 }
